# api/swagger.yaml
basePath: /api
definitions:
...
  endpoints.endpointUpdatePayload:
    properties:
      azureApplicationID:
        description: Azure application ID
        example: eag7cdo9-o09l-9i83-9dO9-f0b23oe78db4
        type: string
      azureAuthenticationKey:
        description: Azure authentication key
        example: cOrXoK/1D35w8YQ8nH1/8ZGwzz45JIYD5jxHKXEQknk=
        type: string
      azureTenantID:
        description: Azure tenant ID
        example: 34ddc78d-4fel-2358-8cc1-df84c8o839f5
        type: string
      edgeCheckinInterval:
        description: The check in interval for edge agent (in seconds)
        example: 5
        type: integer
      groupID:
        description: Group identifier
        example: 1
        type: integer
      kubernetes:
        $ref: '#/definitions/portainer.KubernetesData'
        description: Associated Kubernetes data
      name:
        description: Name that will be used to identify this endpoint
        example: my-endpoint
        type: string
      publicURL:
        description: |-
          URL or IP address where exposed containers will be reachable.\
          Defaults to URL if not specified
        example: docker.mydomain.tld:2375
        type: string
      status:
        description: The status of the endpoint (1 - up, 2 - down)
        example: 1
        type: integer
      tagIDs:
        description: List of tag identifiers to which this endpoint is associated
        example:
        - 1
        - 2
        items:
          description: Tag identifier
          example: 1
          type: integer
        type: array
      teamAccessPolicies:
        $ref: '#/definitions/portainer.TeamAccessPolicies'
      tls:
        description: Require TLS to connect against this endpoint
        example: true
        type: boolean
      tlsskipClientVerify:
        description: Skip client verification when using TLS
        example: false
        type: boolean
      tlsskipVerify:
        description: Skip server verification when using TLS
        example: false
        type: boolean
      url:
        description: URL or IP address of a Docker host
        example: docker.mydomain.tld:2375
        type: string
      userAccessPolicies:
        $ref: '#/definitions/portainer.UserAccessPolicies'
    type: object
...
  portainer.AccessPolicy:
    properties:
      RoleId:
        description: Role identifier. Reference the role that will be associated to
          this access policy
        example: 1
        type: integer
    type: object
...
  portainer.AzureCredentials:
    properties:
      ApplicationID:
        description: Azure application ID
        example: eag7cdo9-o09l-9i83-9dO9-f0b23oe78db4
        type: string
      AuthenticationKey:
        description: Azure authentication key
        example: cOrXoK/1D35w8YQ8nH1/8ZGwzz45JIYD5jxHKXEQknk=
        type: string
      TenantID:
        description: Azure tenant ID
        example: 34ddc78d-4fel-2358-8cc1-df84c8o839f5
        type: string
    type: object
...
  portainer.DockerSnapshot:
    properties:
      DockerSnapshotRaw:
        $ref: '#/definitions/portainer.DockerSnapshotRaw'
      DockerVersion:
        type: string
      HealthyContainerCount:
        type: integer
      ImageCount:
        type: integer
      RunningContainerCount:
        type: integer
      ServiceCount:
        type: integer
      StackCount:
        type: integer
      StoppedContainerCount:
        type: integer
      Swarm:
        type: boolean
      Time:
        type: integer
      TotalCPU:
        type: integer
      TotalMemory:
        type: integer
      UnhealthyContainerCount:
        type: integer
      VolumeCount:
        type: integer
    type: object
  portainer.DockerSnapshotRaw:
    properties:
      Containers:
        type: object
      Images:
        type: object
      Info:
        type: object
      Networks:
        type: object
      Version:
        type: object
      Volumes:
        type: object
    type: object
...
  portainer.Endpoint:
    properties:
      AuthorizedTeams:
        items:
          type: integer
        type: array
      AuthorizedUsers:
        description: Deprecated in DBVersion == 18
        items:
          description: User identifier who created this template
          example: 3
          type: integer
        type: array
      AzureCredentials:
        $ref: '#/definitions/portainer.AzureCredentials'
      EdgeCheckinInterval:
        description: The check in interval for edge agent (in seconds)
        example: 5
        type: integer
      EdgeID:
        description: The identifier of the edge agent associated with this endpoint
        type: string
      EdgeKey:
        description: The key which is used to map the agent to Portainer
        type: string
      Extensions:
        items:
          $ref: '#/definitions/portainer.EndpointExtension'
        type: array
      GroupId:
        description: Endpoint group identifier
        example: 1
        type: integer
      Id:
        description: Endpoint Identifier
        example: 1
        type: integer
      Kubernetes:
        $ref: '#/definitions/portainer.KubernetesData'
        description: Associated Kubernetes data
      Name:
        description: Endpoint name
        example: my-endpoint
        type: string
      PublicURL:
        description: URL or IP address where exposed containers will be reachable
        example: docker.mydomain.tld:2375
        type: string
      Snapshots:
        description: List of snapshots
        items:
          $ref: '#/definitions/portainer.DockerSnapshot'
        type: array
      Status:
        description: The status of the endpoint (1 - up, 2 - down)
        example: 1
        type: integer
      TLS:
        description: |-
          Deprecated fields
          Deprecated in DBVersion == 4
        type: boolean
      TLSCACert:
        type: string
      TLSCert:
        type: string
      TLSConfig:
        $ref: '#/definitions/portainer.TLSConfiguration'
      TLSKey:
        type: string
      TagIds:
        description: List of tag identifiers to which this endpoint is associated
        items:
          description: Tag identifier
          example: 1
          type: integer
        type: array
      Tags:
        description: Deprecated in DBVersion == 22
        items:
          type: string
        type: array
      TeamAccessPolicies:
        $ref: '#/definitions/portainer.TeamAccessPolicies'
        description: List of team identifiers authorized to connect to this endpoint
      Type:
        description: Endpoint environment type. 1 for a Docker environment, 2 for
          an agent on Docker environment or 3 for an Azure environment.
        example: 1
        type: integer
      URL:
        description: URL or IP address of the Docker host associated to this endpoint
        example: docker.mydomain.tld:2375
        type: string
      UserAccessPolicies:
        $ref: '#/definitions/portainer.UserAccessPolicies'
        description: List of user identifiers authorized to connect to this endpoint
    type: object
...
  portainer.EndpointExtension:
    properties:
      Type:
        type: integer
      URL:
        type: string
    type: object
...
  portainer.KubernetesConfiguration:
    properties:
      IngressClasses:
        items:
          $ref: '#/definitions/portainer.KubernetesIngressClassConfig'
        type: array
      StorageClasses:
        items:
          $ref: '#/definitions/portainer.KubernetesStorageClassConfig'
        type: array
      UseLoadBalancer:
        type: boolean
      UseServerMetrics:
        type: boolean
      RestrictDefaultNamespace:
        type: boolean
    type: object
  portainer.KubernetesData:
    properties:
      Configuration:
        $ref: '#/definitions/portainer.KubernetesConfiguration'
      Snapshots:
        items:
          $ref: '#/definitions/portainer.KubernetesSnapshot'
        type: array
    type: object
  portainer.KubernetesIngressClassConfig:
    properties:
      Name:
        type: string
      Type:
        type: string
    type: object
  portainer.KubernetesSnapshot:
    properties:
      KubernetesVersion:
        type: string
      NodeCount:
        type: integer
      Time:
        type: integer
      TotalCPU:
        type: integer
      TotalMemory:
        type: integer
    type: object
  portainer.KubernetesStorageClassConfig:
    properties:
      AccessModes:
        items:
          type: string
        type: array
      AllowVolumeExpansion:
        type: boolean
      Name:
        type: string
      Provisioner:
        type: string
    type: object
...
  portainer.TLSConfiguration:
    properties:
      TLS:
        description: Use TLS
        example: true
        type: boolean
      TLSCACert:
        description: Path to the TLS CA certificate file
        example: /data/tls/ca.pem
        type: string
      TLSCert:
        description: Path to the TLS client certificate file
        example: /data/tls/cert.pem
        type: string
      TLSKey:
        description: Path to the TLS client key file
        example: /data/tls/key.pem
        type: string
      TLSSkipVerify:
        description: Skip the verification of the server TLS certificate
        example: false
        type: boolean
    type: object
...
  portainer.TeamAccessPolicies:
    additionalProperties:
      $ref: '#/definitions/portainer.AccessPolicy'
    type: object
...
  portainer.UserAccessPolicies:
    additionalProperties:
      $ref: '#/definitions/portainer.AccessPolicy'
    type: object
...
info:
  contact:
    email: info@portainer.io
  description: |
    Portainer API is an HTTP API served by Portainer. It is used by the Portainer UI and everything you can do with the UI can be done using the HTTP API.
    Examples are available at https://gist.github.com/deviantony/77026d402366b4b43fa5918d41bc42f8
    You can find out more about Portainer at [http://portainer.io](http://portainer.io) and get some support on [Slack](http://portainer.io/slack/).

    # Authentication

    Most of the API endpoints require to be authenticated as well as some level of authorization to be used.
    Portainer API uses JSON Web Token to manage authentication and thus requires you to provide a token in the **Authorization** header of each request
    with the **Bearer** authentication mechanism.

    Example:

    ```
    Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwidXNlcm5hbWUiOiJhZG1pbiIsInJvbGUiOjEsImV4cCI6MTQ5OTM3NjE1NH0.NJ6vE8FY1WG6jsRQzfMqeatJ4vh2TWAeeYfDhP71YEE
    ```

    # Security

    Each API endpoint has an associated access policy, it is documented in the description of each endpoint.

    Different access policies are available:

    - Public access
    - Authenticated access
    - Restricted access
    - Administrator access

    ### Public access

    No authentication is required to access the endpoints with this access policy.

    ### Authenticated access

    Authentication is required to access the endpoints with this access policy.

...

...
paths:
...
  /endpoint_groups/{id}/endpoints/{endpointId}:
    delete:
      description: '**Access policy**: administrator'
      operationId: EndpointGroupDeleteEndpoint
      parameters:
      - description: EndpointGroup identifier
        in: path
        name: id
        required: true
        type: integer
      - description: Endpoint identifier
        in: path
        name: endpointId
        required: true
        type: integer
      responses:
        "204":
          description: Success
        "400":
          description: Invalid request
        "404":
          description: EndpointGroup not found
        "500":
          description: Server error
      security:
      - jwt: []
      summary: Removes endpoint from an endpoint group
      tags:
      - endpoint_groups
    put:
      description: |-
        Add an endpoint to an endpoint group
        **Access policy**: administrator
      operationId: EndpointGroupAddEndpoint
      parameters:
      - description: EndpointGroup identifier
        in: path
        name: id
        required: true
        type: integer
      - description: Endpoint identifier
        in: path
        name: endpointId
        required: true
        type: integer
      responses:
        "204":
          description: Success
        "400":
          description: Invalid request
        "404":
          description: EndpointGroup not found
        "500":
          description: Server error
      security:
      - jwt: []
      summary: Add an endpoint to an endpoint group
      tags:
      - endpoint_groups
  /endpoints:
    get:
      description: |-
        List all endpoints based on the current user authorizations. Will
        return all endpoints if using an administrator account otherwise it will
        only return authorized endpoints.
        **Access policy**: restricted
      operationId: EndpointList
      parameters:
      - description: Start searching from
        in: query
        name: start
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      - description: List endpoints of this group
        in: query
        name: groupId
        type: integer
      - description: Limit results to this value
        in: query
        name: limit
        type: integer
      - description: List endpoints of this type
        in: query
        name: type
        type: integer
      - description: search endpoints with these tags (depends on tagsPartialMatch)
        in: query
        items:
          type: integer
        name: tagIds
        type: array
      - description: If true, will return endpoint which has one of tagIds, if false
          (or missing) will return only endpoints that has all the tags
        in: query
        name: tagsPartialMatch
        type: boolean
      - description: will return only these endpoints
        in: query
        items:
          type: integer
        name: endpointIds
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: Endpoints
          schema:
            items:
              $ref: '#/definitions/portainer.Endpoint'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: Server
      security:
      - jwt: []
      summary: List endpoints
      tags:
      - endpoints
    post:
      consumes:
      - multipart/form-data
      description: |-
        Create a new endpoint that will be used to manage an environment.
        **Access policy**: administrator
      operationId: EndpointCreate
      parameters:
      - description: 'Name that will be used to identify this endpoint (example: my-endpoint)'
        in: formData
        name: Name
        required: true
        type: string
      - description: 'Environment type. Value must be one of: 1 (Local Docker environment),
          2 (Agent environment), 3 (Azure environment), 4 (Edge agent environment)
          or 5 (Local Kubernetes Environment'
        in: formData
        name: EndpointType
        required: true
        type: integer
      - description: 'URL or IP address of a Docker host (example: docker.mydomain.tld:2375).
          Defaults to local if not specified (Linux: /var/run/docker.sock, Windows:
          //./pipe/docker_engine)'
        in: formData
        name: URL
        type: string
      - description: 'URL or IP address where exposed containers will be reachable.
          Defaults to URL if not specified (example: docker.mydomain.tld:2375)'
        in: formData
        name: PublicURL
        type: string
      - description: Endpoint group identifier. If not specified will default to 1
          (unassigned).
        in: formData
        name: GroupID
        type: integer
      - description: Require TLS to connect against this endpoint
        in: formData
        name: TLS
        type: boolean
      - description: Skip server verification when using TLS
        in: formData
        name: TLSSkipVerify
        type: boolean
      - description: Skip client verification when using TLS
        in: formData
        name: TLSSkipClientVerify
        type: boolean
      - description: TLS CA certificate file
        in: formData
        name: TLSCACertFile
        type: file
      - description: TLS client certificate file
        in: formData
        name: TLSCertFile
        type: file
      - description: TLS client key file
        in: formData
        name: TLSKeyFile
        type: file
      - description: Azure application ID. Required if endpoint type is set to 3
        in: formData
        name: AzureApplicationID
        type: string
      - description: Azure tenant ID. Required if endpoint type is set to 3
        in: formData
        name: AzureTenantID
        type: string
      - description: Azure authentication key. Required if endpoint type is set to
          3
        in: formData
        name: AzureAuthenticationKey
        type: string
      - description: List of tag identifiers to which this endpoint is associated
        in: formData
        items:
          type: integer
        name: TagIDs
        type: array
      - description: The check in interval for edge agent (in seconds)
        in: formData
        name: EdgeCheckinInterval
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/portainer.Endpoint'
        "400":
          description: Invalid request
        "500":
          description: Server error
      security:
      - jwt: []
      summary: Create a new endpoint
      tags:
      - endpoints
  /endpoints/{id}:
    delete:
      description: |-
        Remove an endpoint.
        **Access policy**: administrator
      operationId: EndpointDelete
      parameters:
      - description: Endpoint identifier
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: Success
        "400":
          description: Invalid request
        "404":
          description: Endpoint not found
        "500":
          description: Server error
      security:
      - jwt: []
      summary: Remove an endpoint
      tags:
      - endpoints
    get:
      description: |-
        Retrieve details about an endpoint.
        **Access policy**: restricted
      operationId: EndpointInspect
      parameters:
      - description: Endpoint identifier
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/portainer.Endpoint'
        "400":
          description: Invalid request
        "404":
          description: Endpoint not found
        "500":
          description: Server error
      security:
      - jwt: []
      summary: Inspect an endpoint
      tags:
      - endpoints
    put:
      consumes:
      - application/json
      description: |-
        Update an endpoint.
        **Access policy**: administrator
      operationId: EndpointUpdate
      parameters:
      - description: Endpoint identifier
        in: path
        name: id
        required: true
        type: integer
      - description: Endpoint details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/endpoints.endpointUpdatePayload'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/portainer.Endpoint'
        "400":
          description: Invalid request
        "404":
          description: Endpoint not found
        "500":
          description: Server error
      security:
      - jwt: []
      summary: Update an endpoint
      tags:
      - endpoints
...
schemes:
...
securityDefinitions:
  jwt:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...